Require Control Panel subscriptions for "push" style outgoing commands
- Control Panel commands initiated by the mod will now only send to clients that have asked for them
- Implemented subscription system that groups commands
- Outgoing commands that are only sent in response to client-initiated commands are enabled by default
- Defer message serialization until the system knows there's at least one client to send it to (should improve performance when no clients are connected)
- Not sending anything should avoid interactions with antivirus software
- Migrate Control Panel page to use the H2HWebsocketCore

This should mean the game will only ever try to write anything to the socket when there is an active control panel client. All changes should be backwards-compatible but the old version of the page will not get all the info from the new version of the mod. That'll be okay though.